+++ .eslintrc.js
@@ -1,4 +1,8 @@
 module.exports = {
   root: true,
   extends: '@react-native-community',
+  rules: {
+    'no-shadow': 'off',
+    '@typescript-eslint/no-shadow': 'error',
+  },
 };
