+++ src/screens/staking/utils.ts
@@ -1,0 +1,5 @@
+export enum EViews {
+  info,
+  rewards,
+  history,
+}
